requirements.txt
opengsq>=3.3.0
discord-webhook>=1.3.0
SQLAlchemy>=2.0.27
python-dotenv>=1.0.1
pyyaml>=6.0.1
schedule>=1.2.1
colorlog>=6.8.2
netifaces>=0.11.0 